uppfærði ýmislegt, breytti pom skrá , bætti login ferli , inn og utskraning , vildi ekki over-writa ykkar , gerði new branch

+++ src/main/java/is/hi/hbv501g/moviesuggestor/Controllers/HomeController.java
@@ -4,10 +4,10 @@ import is.hi.hbv501g.moviesuggestor.Persistence.Entities.Movie;
 import is.hi.hbv501g.moviesuggestor.Persistence.Entities.User;
 import is.hi.hbv501g.moviesuggestor.Services.MovieService;
 import is.hi.hbv501g.moviesuggestor.Services.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,53 +17,65 @@ import java.util.List;
 
 @Controller
 public class HomeController {
-    private MovieService movieService;
-    private UserService userService;
+
+    private final MovieService movieService;
+    private final UserService userService;
 
     @Autowired
-    public HomeController(MovieService movieService) {
+    public HomeController(MovieService movieService, UserService userService) {
         this.movieService = movieService;
-    }
-
-    @Autowired
-    public void setUserService(UserService userService) {
         this.userService = userService;
     }
 
+    // Heimasíða
     @RequestMapping("/")
-    public String homePage(Model model) {
-        //Call a method in Service Class
-        List<Movie> allMovies =movieService.findAllMovies();
-
-        List<User> allUsers =userService.findAllUsers();
-        model.addAttribute("users", allUsers);
-
-        //Add some data to the model
+    public String homePage(Model model, HttpSession session) {
+        List<Movie> allMovies = movieService.findAllMovies();
         model.addAttribute("movies", allMovies);
+
+        User loggedInUser = (User) session.getAttribute("LoggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInMessage", "You are logged in as: " + loggedInUser.getUsername());
+            model.addAttribute("isLoggedIn", true); // Senda upplýsingum til að sýna rétt hnapp
+        } else {
+            model.addAttribute("loggedInMessage", "You are not logged in, welcome Guest.");
+            model.addAttribute("isLoggedIn", false); // Senda upplýsingum til að sýna rétt hnapp
+        }
+
         return "home";
     }
 
-    @RequestMapping(value = "/addmovie",method = RequestMethod.GET)
-    public String addmovie(Movie movie) {
-
+    // Ný mynd form
+    @RequestMapping(value = "/addmovie", method = RequestMethod.GET)
+    public String addMovieForm(Movie movie) {
         return "newMovie";
     }
 
-    @RequestMapping(value = "/addmovie",method = RequestMethod.POST)
-    public String addMovie(Movie movie, BindingResult result, Model model) {
-        if(result.hasErrors()) {
-            return "newMovie";
-        }
+    // Bæta við mynd
+    @RequestMapping(value = "/addmovie", method = RequestMethod.POST)
+    public String addMovie(@ModelAttribute Movie movie) {
         movieService.saveMovie(movie);
         return "redirect:/";
     }
 
+    // Eyða mynd
+    @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
+    public String deleteMovie(@PathVariable("id") long id, Model model) {
+        Movie movieToDelete = movieService.findMovieById(id);
+        if (movieToDelete == null) {
+            model.addAttribute("error", "Movie not found");
+            return "error";
+        }
 
-
-    @RequestMapping(value = "/delete/{id}",method = RequestMethod.GET)
-    public String deleteMovie(@PathVariable ("id") long id, Model model) {
-        Movie movieToDelete=movieService.findMovieById(id);
         movieService.deleteMovie(movieToDelete);
         return "redirect:/";
+    }
+
+    // Sýna allar myndir
+    @RequestMapping(value = "/movie", method = RequestMethod.GET)
+    public String getAllMovies(Model model) {
+        List<Movie> allMovies = movieService.findAllMovies();
+        model.addAttribute("movies", allMovies);
+        return "movieList";
     }
 }
